feat: implementacion RegistroPesajeService y persistencia Mongo

+++ src/main/java/cl/sansaweigh/sansaweighapp/dto/RegistroPesajeResponse.java
@@ -21,7 +21,7 @@ public class RegistroPesajeResponse {
 
     private String id;
 
-    private Long balanzaId;
+    private int balanzaId;
 
     private String paqueteId;
